Fine tuning template per Mory:
The line break below the logo was designed for the hunkerhaul treatment if we were to have it under the "hunker" logo. In this case, we could actually remove the line break and decrease the 30px padding we have, and bump the hunker logo to 85x37?

Could we up the letter spacing for the CTA's from 0.5px to 1px and decrease the padding between the header / subcopy to 15px? Then center the hero copy? So we  would have something that looks like the attached image above.